+++ README.md
@@ -1,16 +1,17 @@
 # PO-Project
 
 ## TO DO
-1. rozdzielić klasy do różnych plików
+1. rozdzielić klasy do różnych plików - done
 2. zrobić system zbierania i zapisu danych
 3. zrobić klasy jednostek specjalnych
-4. napisać metodę walki i otrzymywania obrażeń
+4. napisać metodę walki i otrzymywania obrażeń - in progress
 5. napisać warunek końca symulacji
 6. doprcować konstruktor metody Simulation
-7. zrobić komentarze w kodzie co by się później nie pogubić
+7. zrobić komentarze w kodzie co by się później nie pogubić - in progress
 
 
 Jędrzej - Diagram przypadków użycia, karty CRC
+
 Mikołaj - Diagram klas i obiektów
 
 https://drive.google.com/drive/folders/1oXZvaAFNl-0PXs05L2YzsMoFZCARuWrg?usp=sharing
