+++ army.py
@@ -1,6 +1,4 @@
-import config
-import random
-from board import Board
+from unit import Base_unit
 
 class Army:
     def __init__(self, fraction, init_number_of_base_units, init_number_of_special_units, start_point_x, start_point_y):
@@ -34,38 +32,3 @@ class Army:
 
 
 
-class Base_unit:
-    def __init__(self, id, fraction, pos_x, pos_y):
-        self.__id = id
-        self.__fraction = fraction
-        self.__health_points = config.BASE_UNIT_HP
-        self.__strength = config.BASE_UNIT_STRENGTH
-        self.__movement_points = config.BASE_UNIT_MOVMENT_POINT
-        self.__pos_x = pos_x
-        self.__pos_y = pos_y
-        self.__is_alive = True
-
-    def move(self, board):
-        if self.__is_alive == True:
-            movement = self.__movement_points
-            while movement > 0:
-                rand_x = random.randint(-1, 1)
-                rand_y = random.randint(-1, 1)
-                pos_x = self.__pos_x + rand_x
-                pos_y = self.__pos_y + rand_y
-
-                move = board.is_it_free(pos_x, pos_y)
-                if move == 1:
-                    self.capture_the_field()
-                    movement -= 1
-                if move == 2:
-                    self.fight()
-    
-    def capture_the_field(self):
-        pass
-
-    def fight(self):
-        pass
-
-    def die(self):
-        self.__is_alive = False
